feat: moved deleteLifecycleRules inside bucketInfo
