+++ google-cloud-storage/clirr-ignored-differences.xml
@@ -21,9 +21,4 @@
         <method>com.google.cloud.storage.PostPolicyV4 generateSignedPostPolicyV4(com.google.cloud.storage.BlobInfo, long, java.util.concurrent.TimeUnit, com.google.cloud.storage.Storage$PostPolicyV4Option[])</method>
         <differenceType>7012</differenceType>
     </difference>
-    <difference>
-        <className>com/google/cloud/storage/Storage</className>
-        <method>java.util.List deleteLifecycleRules(java.lang.String, com.google.cloud.storage.BucketInfo$LifecycleRule[])</method>
-        <differenceType>7012</differenceType>
-    </difference>
 </differences>
