+++ google-cloud-storage/src/main/java/com/google/cloud/storage/Bucket.java
@@ -1115,45 +1115,6 @@ public class Bucket extends BucketInfo {
   }
 
   /**
-   * Deletes the lifecycle rules of this bucket.
-   *
-   * <p>Example of deleting the lifecycle rules of this bucket:
-   *
-   * <pre>{@code
-   * String bucketName = "my-unique-bucket";
-   * LifecycleRule lifecycleRule_1 =
-   * 	new LifecycleRule(
-   * 		LifecycleAction.newSetStorageClassAction(StorageClass.COLDLINE),
-   * 		LifecycleCondition.newBuilder()
-   * 			.setAge(1)
-   * 			.setNumberOfNewerVersions(3)
-   * 			.setIsLive(false)
-   * 			.setMatchesStorageClass(ImmutableList.of(StorageClass.COLDLINE))
-   * 			.build());
-   * LifecycleRule lifecycleRule_2 =
-   * 	new LifecycleRule(
-   * 		LifecycleAction.newDeleteAction(), LifecycleCondition.newBuilder().setAge(1).build());
-   * ImmutableList<LifecycleRule> lifecycleRules =
-   * 	ImmutableList.of(lifecycleRule_1, lifecycleRule_2);
-   * Bucket bucket =
-   * 	storage.create(
-   * 		BucketInfo.newBuilder(bucketName)
-   * 			.setLocation("us")
-   * 			.setLifecycleRules(lifecycleRules)
-   * 			.build());
-   * Map<LifecycleRule, Boolean> results = storage.deleteLifecycleRules(lifecycleTestBucket, lifecycleRule_1);
-   * }</pre>
-   *
-   * @param rulesToDelete the set of lifecycle rules to delete
-   * @return the lists of deleted lifecycle rules of bucket, an empty list if the requested
-   *     lifecycle rules was not found
-   * @throws StorageException upon failure
-   */
-  public List<LifecycleRule> deleteLifecycleRules(LifecycleRule... rulesToDelete) {
-    return storage.deleteLifecycleRules(getName(), rulesToDelete);
-  }
-
-  /**
    * Creates a new default blob ACL entry on this bucket.
    *
    * <p>Default ACLs are applied to a new blob within the bucket when no ACL was provided for that
